Add struct for character
We need a character that can track its orientation and position. But I
cannot subclass Character as it's closed for sub-classing outside its
module. So I create a struct (Scout) with properties for Character,
Coordinate and Direction.

Add an initializer and a function to materialize the scout character
in the visual world