Add isReallyBlocked() and turnLeft() to scout
We need to know when our scout is front blocked.
But we cannot use Character's isBlocked because the character can jump
over one level gradient (for which isBlocked would return true). So we
compare difference of height between current position and the aimed at
position. The easy case of front blocking is when current position is
at the edge and the scout is facing outward. I've added the turnLeft()
for the test (but we will need it later too).