Fix the edge case for blocking detection
The simple case of detecting is scout is blocked whereby the scout is
already on the edge and looking outward was broken or more precisely
forgot to change it in the last commit. So instead of having one edge
detection for each cardinal points, create a generic edge detection
that take a Direction. In the process of doing so, I realized, the
code in each case of the switch block of isReallyBlocked() was
identical, so I refactored it (effectively the need for that switch
block was made redundant with the gaze() method).