Terrain is generated procedurally using Perlin Noise algorhitm
We use GameplayKit's noise functions to generate a random heightmap
using the Perlin noise algorithm (for smoother realistic edges).
Therefore, we retire the code, the height data for making the valley
and the old height map structure. I've also removed the assertions
that were made for the old landscape. Also already noticed some new
bugs to test for.